// frontend/src/app/api/generated/model/receipt-dto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MoneyDto } from './money-dto';


export interface ReceiptDto { 
    receiptNo?: string;
    patientUid?: string;
    amount?: MoneyDto;
    paymentMode?: string;
    status?: string;
    cashier?: string;
    businessDayUid?: string;
    businessDate?: string;
}